+++ db/migrate/20250407175609_add_diy_kit_to_products_and_product_variants.rb
@@ -1,0 +1,6 @@
+class AddDiyKitToProductsAndProductVariants < ActiveRecord::Migration[8.0]
+  def change
+    add_column :products, :diy_kit, :boolean
+    add_column :product_variants, :diy_kit, :boolean
+  end
+end